deprecation: move `types.d.ts` to `std/io`
