+++ _tools/check_circular_submodule_dependencies.ts
@@ -38,7 +38,7 @@ async function check(
     }
   }
   deps.delete(submod);
-  deps.delete("types.d.ts");
+  deps.delete("io/types.d.ts");
   return { name: submod, set: deps, state };
 }
