+++ archive/_common.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
 import { PartialReadError } from "../io/buf_reader.ts";
-import type { Reader } from "../types.d.ts";
+import type { Reader } from "../io/types.d.ts";
 
 export interface TarInfo {
   fileMode?: number;
