+++ io/buf_writer_test.ts
@@ -7,7 +7,7 @@ import { BufWriter, BufWriterSync } from "./buf_writer.ts";
 import { Buffer } from "./buffer.ts";
 import { StringWriter } from "./string_writer.ts";
 import { bufsizes } from "./_test_common.ts";
-import type { Writer, WriterSync } from "../types.d.ts";
+import type { Writer, WriterSync } from "./types.d.ts";
 
 Deno.test("bufioWriter", async function () {
   const data = new Uint8Array(8192);
