+++ io/read_delim.ts
@@ -2,7 +2,7 @@
 // This module is browser compatible.
 
 import { concat } from "../bytes/concat.ts";
-import type { Reader } from "../types.d.ts";
+import type { Reader } from "./types.d.ts";
 
 /** Generate longest proper prefix which is also suffix array. */
 function createLPS(pat: Uint8Array): Uint8Array {
